+++ wordpress/docs/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to the **wordpress** extension will be documented in this file.
 
+## [3.35.2] - 2026-08-14
+
+### Changed
+- Bound WordPress timeout diagnostics
+
+## [3.35.1] - 2026-08-14
+
+### Fixed
+- Fix WordPress inventory stdout contract
+
 ## [3.35.0] - 2026-08-14
 
 ### Added